requirements.txt
praw~=7.2.0
python-dotenv~=0.15.0